Add React ESLint configuration
