+++ packages/css-in-js/package.json
@@ -47,8 +47,8 @@
     "@rollup/plugin-typescript": "^8.1.0",
     "@types/jest": "^26.0.22",
     "@types/stylis": "^4.0.0",
-    "documentation": "^13.1.0",
-    "eslint": "^7.16.0",
+    "documentation": "^13.2.1",
+    "eslint": "^7.24.0",
     "jest": "^26.6.3",
     "lint-staged": "^10.5.4",
     "prettier": "^2.2.1",
@@ -62,7 +62,7 @@
   },
   "dependencies": {
     "@emotion/hash": "^0.8.0",
-    "@rocket.chat/memo": "^0.23.0",
+    "@rocket.chat/memo": "workspace:packages/memo",
     "stylis": "^4.0.6"
   }
 }
