+++ packages/emitter/package.json
@@ -52,8 +52,8 @@
     "@rollup/plugin-node-resolve": "^11.0.1",
     "@rollup/plugin-typescript": "^8.1.0",
     "@types/jest": "^26.0.22",
-    "documentation": "^13.1.0",
-    "eslint": "^7.16.0",
+    "documentation": "^13.2.1",
+    "eslint": "^7.24.0",
     "jest": "^26.6.3",
     "lint-staged": "^10.5.4",
     "npm-run-all": "^4.1.5",
