+++ packages/eslint-config-alt/package.json
@@ -19,7 +19,9 @@
   "main": "minimal/index.js",
   "files": [
     "/minimal",
-    "/original"
+    "/original",
+    "/react",
+    "/typescript"
   ],
   "publishConfig": {
     "access": "public"
@@ -44,8 +46,12 @@
   },
   "dependencies": {
     "@rocket.chat/eslint-config": "^0.4.0",
+    "@typescript-eslint/eslint-plugin": "^4.22.0",
+    "@typescript-eslint/parser": "^4.22.0",
     "eslint-config-prettier": "^8.2.0",
     "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-prettier": "^3.4.0"
+    "eslint-plugin-prettier": "^3.4.0",
+    "eslint-plugin-react": "^7.23.2",
+    "eslint-plugin-react-hooks": "^4.2.0"
   }
 }
