+++ packages/fuselage-hooks/package.json
@@ -10,7 +10,8 @@
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/RocketChat/Rocket.Chat.Fuselage.git"
+    "url": "git+https://github.com/RocketChat/Rocket.Chat.Fuselage.git",
+    "directory": "packages/fuselage-hooks"
   },
   "bugs": {
     "url": "https://github.com/RocketChat/Rocket.Chat.Fuselage/issues"
@@ -29,7 +30,9 @@
   "files": [
     "/dist"
   ],
-  "prettier": "@rocket.chat/prettier-config/fuselage",
+  "publishConfig": {
+    "access": "public"
+  },
   "scripts": {
     "start": "rollup -c -w",
     "build": "rollup -c",
@@ -47,45 +50,37 @@
   "devDependencies": {
     "@microsoft/api-documenter": "^7.8.21",
     "@microsoft/api-extractor": "^7.9.2",
-    "@rocket.chat/eslint-config": "^0.4.0",
+    "@rocket.chat/eslint-config-alt": "workspace:packages/eslint-config-alt",
+    "@rocket.chat/fuselage-tokens": "workspace:packages/fuselage-tokens",
     "@rocket.chat/prettier-config": "workspace:packages/prettier-config",
     "@rollup/plugin-commonjs": "^15.1.0",
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^9.0.0",
     "@rollup/plugin-typescript": "^6.0.0",
     "@types/jest": "^26.0.22",
-    "@types/react-dom": "^16.9.8",
-    "@types/resize-observer-browser": "^0.1.3",
+    "@types/react": "^17.0.3",
+    "@types/react-dom": "^17.0.3",
+    "@types/resize-observer-browser": "^0.1.5",
     "@types/use-subscription": "^1.0.0",
-    "@typescript-eslint/eslint-plugin": "^4.5.0",
-    "@typescript-eslint/parser": "^4.5.0",
-    "documentation": "^13.0.2",
-    "eslint": "^7.13.0",
-    "eslint-config-prettier": "^6.15.0",
-    "eslint-plugin-import": "^2.19.1",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.16.0",
-    "eslint-plugin-react-hooks": "^4.0.4",
+    "documentation": "^13.2.1",
+    "eslint": "^7.24.0",
     "jest": "^26.0.1",
     "lint-staged": "^10.5.4",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.1.2",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "rimraf": "^3.0.2",
     "rollup": "^2.21.0",
     "rollup-plugin-terser": "^7.0.2",
     "ts-jest": "^26.1.0",
-    "typescript": "^4.0.3"
+    "typescript": "^4.0.3",
+    "use-subscription": "^1.5.1"
   },
   "peerDependencies": {
-    "react": "^16.12.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {
-    "@rocket.chat/fuselage-tokens": "^0.23.0",
-    "use-subscription": "^1.4.1"
+    "@rocket.chat/fuselage-tokens": "*",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "use-subscription": "^1.5.1"
   }
 }
