+++ packages/fuselage-ui-kit/.eslintrc.js
@@ -1,59 +1,3 @@
 module.exports = {
-  extends: ['@rocket.chat/eslint-config-alt/original', 'prettier'],
-  plugins: ['react', 'react-hooks', 'prettier'],
-  parser: '@babel/eslint-parser',
-  rules: {
-    'generator-star-spacing': ['error', 'before'],
-    'import/order': [
-      'error',
-      {
-        'newlines-between': 'always',
-        'groups': [
-          'builtin',
-          'external',
-          'internal',
-          ['parent', 'sibling', 'index'],
-        ],
-        'alphabetize': {
-          order: 'asc',
-        },
-      },
-    ],
-    'indent': 'off',
-    'jsx-quotes': ['error', 'prefer-single'],
-    'react/jsx-uses-react': 'error',
-    'react/jsx-uses-vars': 'error',
-    'react/jsx-no-undef': 'error',
-    'react/jsx-fragments': ['error', 'syntax'],
-    'react/react-in-jsx-scope': 'error',
-    'react-hooks/rules-of-hooks': 'error',
-    'react-hooks/exhaustive-deps': 'warn',
-    'prettier/prettier': 2,
-  },
-  settings: {
-    'import/resolver': {
-      node: {
-        extensions: ['.js', '.ts', '.tsx'],
-      },
-    },
-    'react': {
-      version: 'detect',
-    },
-  },
-  env: {
-    jest: true,
-  },
-  overrides: [
-    {
-      files: ['*.mdx'],
-      extends: ['plugin:mdx/recommended'],
-      parserOptions: {
-        parser: '@babel/eslint-parser',
-      },
-      rules: {
-        'semi': 'off',
-        'new-cap': 'off',
-      },
-    },
-  ],
+  extends: '@rocket.chat/eslint-config-alt/react',
 };
