+++ packages/fuselage-ui-kit/src/blocks/ActionsBlock/ActionsBlock.js
@@ -1,31 +1,9 @@
 import { Box, Button } from '@rocket.chat/fuselage';
-import { BlockContext, ElementType } from '@rocket.chat/ui-kit';
+import { BlockContext } from '@rocket.chat/ui-kit';
 import React, { memo, useCallback, useMemo, useState } from 'react';
 
-import { useSurfaceType } from '../surfaces/SurfaceContext';
-
-const Action = ({ element, parser }) => {
-  const renderedElement = parser.renderActions(
-    element,
-    BlockContext.ACTION,
-    parser
-  );
-
-  if (!renderedElement) {
-    return null;
-  }
-
-  return (
-    <Box
-      display='flex'
-      margin={4}
-      flexGrow={element.type !== ElementType.BUTTON ? 1 : undefined}
-      flexBasis={element.type !== ElementType.BUTTON ? '45%' : undefined}
-    >
-      {renderedElement}
-    </Box>
-  );
-};
+import { useSurfaceType } from '../../surfaces/SurfaceContext';
+import Action from './Action';
 
 const ActionsBlock = ({ className, blockElement, parser }) => {
   const surfaceType = useSurfaceType();
