+++ packages/fuselage-ui-kit/src/blocks/ActionsBlock/index.js
@@ -1,0 +1,1 @@
+export { default } from './ActionsBlock';
