+++ packages/fuselage-ui-kit/src/blocks/SectionBlock/Accessory.js
@@ -1,0 +1,10 @@
+import { BlockContext } from '@rocket.chat/ui-kit';
+
+const Accessory = ({ blockId, appId, element, parser }) =>
+  parser.renderAccessories(
+    { blockId, appId, ...element },
+    BlockContext.SECTION,
+    parser
+  );
+
+export default Accessory;
