+++ packages/fuselage-ui-kit/src/blocks/SectionBlock/index.js
@@ -1,0 +1,1 @@
+export { default } from './SectionBlock';
