+++ packages/fuselage-ui-kit/src/surfaces/BannerSurface.js
@@ -3,12 +3,10 @@ import React from 'react';
 
 import { Surface } from './SurfaceContext';
 
-const BannerSurface = ({ children }) => {
-  return (
-    <Surface type='banner'>
-      <Margins block='x8'>{children}</Margins>
-    </Surface>
-  );
-};
+const BannerSurface = ({ children }) => (
+  <Surface type='banner'>
+    <Margins block='x8'>{children}</Margins>
+  </Surface>
+);
 
 export default BannerSurface;
