+++ packages/fuselage-ui-kit/src/surfaces/MessageSurface.js
@@ -3,12 +3,10 @@ import React from 'react';
 
 import { Surface } from './SurfaceContext';
 
-const MessageSurface = ({ children }) => {
-  return (
-    <Surface type='message'>
-      <Margins blockEnd='x16'>{children}</Margins>
-    </Surface>
-  );
-};
+const MessageSurface = ({ children }) => (
+  <Surface type='message'>
+    <Margins blockEnd='x16'>{children}</Margins>
+  </Surface>
+);
 
 export default MessageSurface;
