+++ packages/fuselage-ui-kit/src/surfaces/ModalSurface.js
@@ -3,12 +3,10 @@ import React from 'react';
 
 import { Surface } from './SurfaceContext';
 
-const ModalSurface = ({ children }) => {
-  return (
-    <Surface type='modal'>
-      <Margins blockEnd='x16'>{children}</Margins>
-    </Surface>
-  );
-};
+const ModalSurface = ({ children }) => (
+  <Surface type='modal'>
+    <Margins blockEnd='x16'>{children}</Margins>
+  </Surface>
+);
 
 export default ModalSurface;
