+++ packages/fuselage/.eslintrc.js
@@ -1,50 +1,9 @@
 module.exports = {
-  extends: ['@rocket.chat/eslint-config', 'prettier'],
-  plugins: ['react', 'react-hooks', 'prettier'],
-  parser: '@babel/eslint-parser',
+  extends: '@rocket.chat/eslint-config-alt/react',
   rules: {
-    'generator-star-spacing': ['error', 'before'],
-    'import/order': [
-      'error',
-      {
-        'newlines-between': 'always',
-        'groups': [
-          'builtin',
-          'external',
-          'internal',
-          ['parent', 'sibling', 'index'],
-        ],
-        'alphabetize': {
-          order: 'asc',
-        },
-      },
-    ],
-    'indent': [
-      'error',
-      2,
-      {
-        SwitchCase: 1,
-      },
-    ],
-    'jsx-quotes': ['error', 'prefer-single'],
-    'react/jsx-uses-react': 'error',
-    'react/jsx-uses-vars': 'error',
-    'react/jsx-no-undef': 'error',
-    'react/jsx-fragments': ['error', 'syntax'],
-    'react/react-in-jsx-scope': 'error',
-    'react-hooks/rules-of-hooks': 'error',
-    'react-hooks/exhaustive-deps': 'warn',
-    'prettier/prettier': 2,
-  },
-  settings: {
-    'import/resolver': {
-      node: {
-        extensions: ['.js', '.jsx', '.ts', '.tsx'],
-      },
-    },
-    'react': {
-      version: 'detect',
-    },
+    '@typescript-eslint/explicit-module-boundary-types': 'off',
+    'react/display-name': 'off',
+    'react/no-multi-comp': 'off',
   },
   env: {
     jest: true,
@@ -57,8 +16,9 @@ module.exports = {
         parser: '@babel/eslint-parser',
       },
       rules: {
-        'semi': 'off',
         'new-cap': 'off',
+        'prefer-arrow-callback': 'off',
+        'semi': 'off',
       },
     },
   ],
