+++ packages/fuselage/package.json
@@ -9,7 +9,8 @@
   "homepage": "https://github.com/RocketChat/Rocket.Chat.Fuselage#readme",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/RocketChat/Rocket.Chat.Fuselage.git"
+    "url": "git+https://github.com/RocketChat/Rocket.Chat.Fuselage.git",
+    "directory": "packages/fuselage"
   },
   "bugs": {
     "url": "https://github.com/RocketChat/Rocket.Chat.Fuselage/issues"
@@ -21,7 +22,9 @@
     "/fuselage.d.ts",
     "/index.js"
   ],
-  "prettier": "@rocket.chat/prettier-config/fuselage",
+  "publishConfig": {
+    "access": "public"
+  },
   "scripts": {
     "start": "webpack --watch --mode development",
     "storybook": "start-storybook -p 6006",
@@ -48,16 +51,17 @@
   },
   "peerDependencies": {
     "@rocket.chat/fuselage-hooks": "*",
+    "@rocket.chat/fuselage-polyfills": "*",
     "@rocket.chat/icons": "*",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1"
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
   },
   "dependencies": {
-    "@rocket.chat/css-in-js": "^0.23.0",
-    "@rocket.chat/fuselage-tokens": "^0.23.0",
-    "@rocket.chat/memo": "^0.6.3-dev.180",
+    "@rocket.chat/css-in-js": "workspace:packages/css-in-js",
+    "@rocket.chat/fuselage-tokens": "workspace:packages/fuselage-tokens",
+    "@rocket.chat/memo": "workspace:packages/memo",
     "invariant": "^2.2.4",
-    "react-keyed-flatten-children": "^1.2.0"
+    "react-keyed-flatten-children": "^1.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.11.5",
@@ -67,8 +71,10 @@
     "@babel/preset-env": "^7.11.5",
     "@babel/preset-react": "^7.10.4",
     "@rocket.chat/eslint-config": "^0.4.0",
-    "@rocket.chat/fuselage-hooks": "^0.23.0",
-    "@rocket.chat/fuselage-polyfills": "^0.23.0",
+    "@rocket.chat/eslint-config-alt": "workspace:packages/eslint-config-alt",
+    "@rocket.chat/fuselage-hooks": "workspace:packages/fuselage-hooks",
+    "@rocket.chat/fuselage-polyfills": "workspace:packages/fuselage-polyfills",
+    "@rocket.chat/icons": "workspace:packages/icons",
     "@rocket.chat/prettier-config": "workspace:packages/prettier-config",
     "@storybook/addon-essentials": "^6.1.11",
     "@storybook/addon-jest": "^6.1.11",
@@ -82,14 +88,10 @@
     "babel-loader": "^8.1.0",
     "caniuse-lite": "^1.0.30001151",
     "cross-env": "^7.0.3",
+    "css-loader": "^5.2.2",
     "cssnano": "^4.1.10",
-    "eslint": "^7.13.0",
-    "eslint-config-prettier": "^6.15.0",
-    "eslint-plugin-import": "^2.22.0",
+    "eslint": "^7.24.0",
     "eslint-plugin-mdx": "^1.8.2",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.20.6",
-    "eslint-plugin-react-hooks": "^4.1.0",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^26.6.1",
     "lint-staged": "^10.5.4",
@@ -104,8 +106,8 @@
     "postcss-svg": "^3.0.0",
     "prettier": "^2.1.2",
     "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-live": "^2.2.2",
     "rimraf": "^3.0.2",
     "sass": "^1.26.10",
@@ -137,9 +139,5 @@
         "preset": "iPhone 7"
       }
     }
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "gitHead": "f8467b8ffb6f90d91b46fbcdd5dca169235916a4"
+  }
 }
