+++ packages/ui-kit/package.json
@@ -43,8 +43,8 @@
     "@rocket.chat/prettier-config": "workspace:packages/prettier-config",
     "@types/jest": "^26.0.22",
     "babel-loader": "^8.1.0",
-    "documentation": "^13.1.0",
-    "eslint": "^7.16.0",
+    "documentation": "^13.2.1",
+    "eslint": "^7.24.0",
     "jest": "^26.6.3",
     "lint-staged": "^10.5.4",
     "npm-run-all": "^4.1.5",
